Реализует TripController
